Model_Context_Protocol/pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.12",
    "httpx>=0.28.1",
    "langchain-aws>=0.2.22",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.59",
    "langchain-mcp-adapters>=0.0.11",
    "langgraph>=0.4.3",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
]
